Fix and restore upgrade tests
* Save refarch used during bootstrap
* Add --version paramter to deploy.py
* Fix case default, "*") replaced by *)
* Fix upgrade script and playbook
* Fix jjb error missing /jenkins/api/ redirection
* Add upgrade tests
* Add pip freeze and rpm -qa to image build output
* Bump to 2.0.0

+++ fetch_roles.sh
@@ -17,9 +17,10 @@
 set -e
 [ -n "$DEBUG" ] && set -x
 [ -n "$PB" ] && PB="--progress-bar" || PB="-s"
-[ -z "$2" ] || SF_REL=$2
 
 . ./role_configrc
+
+[ -z "$1" ] || SF_VER=$1
 
 function die {
     echo "[ERROR]: $1"
@@ -32,6 +33,10 @@ function fetch_sf_roles_prebuilt {
     local temp=$(mktemp -d /tmp/edeploy-check-XXXXX)
     for role in softwarefactory; do
         role=${role}-${SF_VER}
+        if [ -f ${UPSTREAM}/${role}.edeploy ]; then
+            echo "${UPSTREAM}/${role}.edeploy already exists..."
+            continue
+        fi
         curl -s -o ${temp}/${role}.md5 ${SWIFT_SF_URL}/${role}.md5 &> /dev/null || die "Could not fetch ${SWIFT_SF_URL}/${role}.md5"
         # Swift does not return 404 but 'Not Found'
         grep -q 'Not Found' ${temp}/${role}.md5 && die "${role} does not exist upstream"
